remove lines that break the docker-build
This change removes some of the security features  that broke the docker-build. Some of the programs and users/groups were needed in the update process. I am going to use this docker-file merely for testing, especially for testing plugins against new releases, so for my purpose this should be fine.
If you want to run mahara with docker in a productive environment, you should consider to look at this lines again.
A solution might probably be to have an intermediate docker image that installs all php-extensions and other programs needed to run mahara. When that is build and installed, the users, groups and programs can probably be removed because they should not be needed for installing mahara itself.

* php-fpm seems not to be installed at the dockerfile-apache
* some of the users and groups were needed to update and install the php-extensions.  ln for example is needed for the setup of various packages